Refactor, added metrics
